Solution to Check if a Number is Prime or Not

+++ Check if a Number is Prime or Not@@6788dd63490637cfc44e1ba2/code.c
@@ -1,18 +1,24 @@
 #include <stdio.h>
 #include <math.h>
+
 int main() {
     int a, i;
     scanf("%d", &a);
+
     if (a <= 1) {
         printf("Not Prime");
         return 0;
     }
-    for (i = 2; i <= sqrt(a); i++) {
+
+    // Loop from 2 to sqrt(a), cast sqrt(a) to an int
+    for (i = 2; i <= (int)sqrt(a); i++) {
         if (a % i == 0) {
             printf("Not Prime");
             return 0;
         }
     }
+
     printf("Prime");
     return 0;
 }
+
